backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
billiard==3.6.4.0
bitarray==2.6.0
boto3==1.24.41
botocore==1.27.41
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.12.0
Deprecated==1.2.13
Django==3.2.6
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.0
django-filter==22.1
django-redis==5.2.0
django-rest-passwordreset==1.2.1
django-solo==1.2.0
django-timezone-field==5.0
djangorestframework==3.13.1
drf-yasg==1.21.3
eth-abi==2.2.0
eth-account==0.5.8
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
inflection==0.5.1
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.8.0
kombu==5.2.4
lru-dict==1.1.8
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.20.1
psycopg2-binary==2.9.3
pycryptodome==3.15.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
rlp==2.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toolz==0.12.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
uuid==1.30
uWSGI==2.0.20
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.30.0
websockets==9.1
wrapt==1.14.1
yarl==1.7.2
